update c51 & c51_jax tests
