+++ tests/test_classic_control.py
@@ -7,11 +7,3 @@ def test_ppo():
         shell=True,
         check=True,
     )
-
-
-def test_c51():
-    subprocess.run(
-        "python cleanrl/c51.py --learning-starts 200 --total-timesteps 205",
-        shell=True,
-        check=True,
-    )
