+++ tests/test_classic_control_gymnasium.py
@@ -7,3 +7,19 @@ def test_dqn():
         shell=True,
         check=True,
     )
+
+
+def test_c51():
+    subprocess.run(
+        "python cleanrl/c51.py --learning-starts 200 --total-timesteps 205",
+        shell=True,
+        check=True,
+    )
+
+
+def test_c51_eval():
+    subprocess.run(
+        "python cleanrl/c51.py --save-model True --learning-starts 200 --total-timesteps 205",
+        shell=True,
+        check=True,
+    )
