--- tests/test_classic_control_jax.py
@@ -1,9 +1,0 @@
-import subprocess
-
-
-def test_dqn_jax():
-    subprocess.run(
-        "python cleanrl/dqn_jax.py --learning-starts 200 --total-timesteps 205",
-        shell=True,
-        check=True,
-    )
